@asmorkalov 2022-10-13
@zihaomu Could you run perf test with 4.x and your branch and attach results to the PR. ARM and some AMD/Intel CPU performance numbers will be useful.