@zihaomu 2022-10-14
Hi @asmorkalov, the performance test has been updated.