@alalek 2022-10-19
Debug builds (no optimization) are broken: http://pullrequest.opencv.org/buildbot/builders/master_noOCL_noICV_noSSE-lin64-debug/builds/100001